Removed circular references, buildRoute should work
Removed circular references between RouteCell and Route, and between Route and Player. RouteCell is a POJO now instead of extending javafx.Rectangle. Click events on the Rectangle overlay are handled by finding the corresponding routecell in a HashMap<RouteCell, Rectangle>, then looping through all Routes until there is a Route which has the given RouteCell as one of its children. Route no longer has a Player owner, instead the Color value of the rectangle overlay is used to determine if a Route has an owner or not. This check is performed so that Players can't build a Route that has already been built by other Players. Also added empty constructors for Firebase